Refactor button link dispatch
